Fex Donations for admin Model campaging and Update doration two again store donations model and migrate

+++ database/migrations/2026_06_06_165114_create_dorations_table.php
@@ -12,18 +12,20 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('dorations', function (Blueprint $table) {
-            $table->id();
-            $table->foreignId('donor_profile_id')->constrained('donor_profiles')->onDelete('cascade');
-            $table->string('name');
-            $table->integer('amount');
-            $table->enum('payment_method', ['نقد', 'بطاقة مصرفية', 'محفظة موبايل', 'تحويل بنكي', 'حوالة']);
-            $table->enum('cat', ['إطعام', 'مساجد', 'تعليم', 'صحة', 'مياه', 'أيتام']);
-            $table->date('date');
-            $table->enum('status', ['مكتمل', 'ملغي', 'قيد المراجعة'])->default('قيد المراجعة');
-            $table->string('notes')->nullable();
-            $table->timestamps();
-        });
+      if (!Schema::hasTable('dorations')) {
+            Schema::create('dorations', function (Blueprint $table) {
+                $table->id();
+                $table->foreignId('donor_profile_id')->constrained('donor_profiles')->onDelete('cascade');
+                $table->string('name');
+                $table->integer('amount');
+                $table->enum('payment_method', ['نقد', 'بطاقة مصرفية', 'محفظة موبايل', 'تحويل بنكي', 'حوالة']);
+                $table->enum('cat', ['إطعام', 'مساجد', 'تعليم', 'صحة', 'مياه', 'أيتام']);
+                $table->date('date');
+                $table->enum('status', ['مكتمل', 'ملغي', 'قيد المراجعة','pending', 'completed', 'failed', 'refunded'])->default('pending');
+                $table->string('notes')->nullable();
+                $table->timestamps();
+            });
+      }
     }
     /**
      * Reverse the migrations.
